# perf: Prefer datetime.fromisoformat over dateutil.parse in Python 3.11+ 