--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.44.0 [unreleased]
 
+### Performance
+1. [#657](https://github.com/influxdata/influxdb-client-python/pull/657): Prefer datetime.fromisoformat over dateutil.parse in Python 3.11+ 
+
 ## 1.43.0 [2024-05-17]
 
 ### Bug Fixes