## Proposed Changes

This PR changes the behavior of `get_date_helper` to prefer `datetime.datetime.fromisoformat` over `dateutil.parser` in Python 3.11 or higher.
According to https://github.com/closeio/ciso8601/blob/b17984f5b4c6db0ef596b4d31456c8a39ae278ee/why_ciso8601.md, `Since Python 3.11+, the performance of cPython's datetime.fromisoformat is now very good`. I did some quick testing and it seems to be on a par with `ciso8601`:
On Python 3.10 I've got ~37,5s and ~3,5s without and with `ciso8601` installed, respectively. On Python 3.11 I've got ~2,5s and ~2,6s without and with `ciso8601` installed, respectively.

Because `ciso` is no dependency, but an extra, not all users (including me) installed it. With this PR these users will get a better out-of-the-box performance for Python 3.11+ users.

## Checklist

<!-- Checkboxes below this note can be erased if not applicable to your Pull Request. -->

- [x] CHANGELOG.md updated
- [x] Rebased/mergeable
- [ ] A test has been added if appropriate
- [x] `pytest tests` completes successfully
- [x] Commit messages are [conventional](https://www.conventionalcommits.org/en/v1.0.0/)
- [x] Sign [CLA](https://www.influxdata.com/legal/cla/) (if not already signed)